fix: Change RuntimeMinutes type from short? to int? in Title class

+++ Adv. Database/Exercise 3 - Data Insert/ImdbDataInsert/Models/Title.cs
@@ -9,7 +9,7 @@
         public bool IsAdult { get; set; }
         public short? StartYear { get; set; }
         public short? EndYear { get; set; }
-        public short? RuntimeMinutes { get; set; }
+        public int? RuntimeMinutes { get; set; }
 
         public override string ToString()
         {
